% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morphology.R
\name{se_box}
\alias{se_box}
\title{Square structuring element}
\usage{
se_box(radius)
}
\arguments{
\item{radius}{integer >= 0}
}
\value{
a plain matrix, all \code{1}s
}
\description{
\code{radius} cells in every direction (so a \code{2*radius + 1} wide matrix of
all \code{1}s).
}
